Use customRequest, move common functions for images to one file
